Rename ElasticsearchResponseBase
